requirements.txt
fastapi==0.123.0
uvicorn==0.38.0
gunicorn==23.0.0

pydantic==2.12.2
pydantic-settings==2.12.0

supabase==2.24.0
httpx==0.28.1

google-genai==1.52.0

pandas==2.3.3
openpyxl==3.1.5

pdfplumber==0.11.8

python-multipart==0.0.9
httpx==0.26.0
python-dotenv==1.2.1
defusedxml==0.7.1

pytest==8.4.2